refactor(Scripts/BlackTemple): Update the Illidari Council script to the new registration method  (#19744)
refactor(Scripts/BlackTemple): Update Illidari council to new model
